+++ engine/api/golang/kurtosis_engine_api_version/kurtosis_engine_api_version.go
@@ -2,5 +2,5 @@ package kurtosis_engine_api_version
 
 const (
 	// DO NOT UPDATE THIS VALUE! It will get updated automatically during the release process
-	KurtosisEngineApiVersion = "1.2.0"
+	KurtosisEngineApiVersion = "1.2.1"
 )