+++ engine/server/engine/server/engine_server_service.go
@@ -5,6 +5,7 @@ import (
 	"github.com/kurtosis-tech/kurtosis-engine-server/api/golang/kurtosis_engine_api_version"
 	"github.com/kurtosis-tech/kurtosis-engine-server/api/golang/kurtosis_engine_rpc_api_bindings"
 	"github.com/kurtosis-tech/kurtosis-engine-server/server/engine/enclave_manager"
+	"github.com/kurtosis-tech/kurtosis-engine-server/server/engine/kurtosis_engine_server_version"
 	"github.com/palantir/stacktrace"
 	"github.com/sirupsen/logrus"
 	"google.golang.org/protobuf/types/known/emptypb"
@@ -27,6 +28,7 @@ func NewEngineServerService(enclaveManager *enclave_manager.EnclaveManager) *Eng
 func (service *EngineServerService) GetEngineInfo(ctx context.Context, empty *emptypb.Empty) (*kurtosis_engine_rpc_api_bindings.GetEngineInfoResponse, error) {
 	result := &kurtosis_engine_rpc_api_bindings.GetEngineInfoResponse{
 		EngineApiVersion: kurtosis_engine_api_version.KurtosisEngineApiVersion,
+		EngineVersion: kurtosis_engine_server_version.KurtosisEngineServerVersion,
 	}
 	return result, nil
 }