+++ engine/docs/changelog.md
@@ -1,5 +1,9 @@
 # TBD
 
+# 1.2.1
+### Features
+* Added `EngineVersion` field that represents the engine server version in the `GetEngineInfo` response
+
 # 1.2.0
 ### Features
 * The launcher makes a best-effort attempt to pull the latest version of the image being started